Update topUpRequest.gs
made the repeated lines of code into functions so that it can be called without writing the same lines of code. gave comments to how each line of code works. set time delays to each of the send email functions because there seems to be a bug that sends empty emails (emails without attachment or any text) to the email addresses